metadata/stream: add digest reference to KubeVirt image description
In the current design, the KubeVirt image is presumed to be a pull spec
with an image digest.  That implies that the user should reference the
image by digest, when we'd prefer that they reference it by a
stream-specific floating tag.

Define the existing "image" field to contain a pull spec with a floating
tag for the stream.  We should still record the unique identifier of the
current image (as we do for GCP images), so add a "digest-ref" field
which always contains a fully-qualified pull spec with digest.